My changes as of 9/25/20
Added new dark_mode setting to user schema.
Reworked settings.vue, main.js and store.js to hold a temporary variable that can be used to change dark_mode in real time. 
Integrated the two variables to load a switch in the correct position, and created a POST method to change it to the displayed one when the user hits submit.

Moving forward: Will clean this up, code is scattered and not as straight forward/ efficient as I'd like it to be. Will use this new capability to actually make a dark mode in css and the like.